refactor: extract PDF paragraph continuations
